+++ .github/workflows/build-test.yml
@@ -26,7 +26,7 @@ jobs:
     - name: Setup .NET 5
       uses: actions/setup-dotnet@v1
       with:
-        dotnet-version: 5.0.102
+        dotnet-version: 5.0.200
     - name: Build
       run: ./publish.cmd
     - name: Test
