+++ Projects/Server/Network/Packets/OutgoingMobilePackets.cs
@@ -17,7 +17,7 @@ using System;
 using System.Buffers;
 using System.IO;
 using System.Runtime.CompilerServices;
-using Microsoft.Toolkit.HighPerformance.Memory;
+using Microsoft.Toolkit.HighPerformance;
 
 namespace Server.Network
 {
