+++ Projects/Server/Server.csproj
@@ -28,7 +28,7 @@
         <Delete Files="..\..\Distribution\$(AssemblyName).runtimeconfig.json" ContinueOnError="true" />
     </Target>
     <ItemGroup>
-        <PackageReference Include="Microsoft.Toolkit.HighPerformance" Version="7.0.0-preview4" />
+        <PackageReference Include="Microsoft.Toolkit.HighPerformance" Version="7.0.0-preview5" />
         <PackageReference Include="Zlib.Bindings" Version="1.4.0" />
     </ItemGroup>
 </Project>
