+++ Projects/Server/Utilities/Utility.cs
@@ -8,7 +8,7 @@ using System.Net.Sockets;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Xml;
-using Microsoft.Toolkit.HighPerformance.Extensions;
+using Microsoft.Toolkit.HighPerformance;
 using Server.Buffers;
 using Server.Network;
 using Server.Random;
