+++ Projects/UOContent/UOContent.csproj
@@ -29,7 +29,7 @@
             <IncludeInPackage>false</IncludeInPackage>
         </ProjectReference>
         <PackageReference Include="MailKit" Version="2.10.1" />
-        <PackageReference Include="Microsoft.Toolkit.HighPerformance" Version="7.0.0-preview4" />
+        <PackageReference Include="Microsoft.Toolkit.HighPerformance" Version="7.0.0-preview5" />
         <PackageReference Include="Zlib.Bindings" Version="1.4.0" />
         <PackageReference Include="Argon2.Bindings" Version="1.8.0" />
     </ItemGroup>
