+++ azure-pipelines.yml
@@ -18,7 +18,7 @@ jobs:
     displayName: 'Install .NET 5'
     inputs:
       packageType: sdk
-      version: '5.0.102'
+      version: 5.0.200
   - task: NuGetAuthenticate@0
   - script: ./publish.cmd Release win
     displayName: 'Build'
@@ -56,7 +56,7 @@ jobs:
     displayName: 'Install .NET 5'
     inputs:
       packageType: sdk
-      version: '5.0.102'
+      version: 5.0.200
   - task: NuGetAuthenticate@0
   - script: ./publish.cmd Release $(os)
     displayName: 'Build'
